Create RectorRecipeInteractiveProvider service
