+++ packages/rector-generator/src/Command/GenerateCommand.php
@@ -11,20 +11,15 @@ use Rector\RectorGenerator\Config\ConfigFilesystem;
 use Rector\RectorGenerator\Finder\TemplateFinder;
 use Rector\RectorGenerator\Generator\FileGenerator;
 use Rector\RectorGenerator\Guard\OverrideGuard;
-use Rector\RectorGenerator\Provider\NodeTypesProvider;
-use Rector\RectorGenerator\Provider\PackageNamesProvider;
+use Rector\RectorGenerator\Provider\RectorRecipeInteractiveProvider;
 use Rector\RectorGenerator\Provider\RectorRecipeProvider;
-use Rector\RectorGenerator\Provider\SetsListProvider;
 use Rector\RectorGenerator\TemplateVariablesFactory;
 use Rector\RectorGenerator\ValueObject\RectorRecipe;
-use Rector\Set\ValueObject\SetList;
 use Rector\Testing\PHPUnit\StaticPHPUnitEnvironment;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\ChoiceQuestion;
-use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symplify\PackageBuilder\Console\ShellCode;
 use Symplify\SmartFileSystem\SmartFileInfo;
@@ -74,19 +69,9 @@ final class GenerateCommand extends Command
     private $rectorRecipeProvider;
 
     /**
-     * @var PackageNamesProvider
+     * @var RectorRecipeInteractiveProvider
      */
-    private $packageNamesProvider;
-
-    /**
-     * @var NodeTypesProvider
-     */
-    private $nodeTypesProvider;
-
-    /**
-     * @var SetsListProvider
-     */
-    private $setsListProvider;
+    private $rectorRecipeInteractiveProvider;
 
     public function __construct(
         ComposerPackageAutoloadUpdater $composerPackageAutoloadUpdater,
@@ -96,9 +81,7 @@ final class GenerateCommand extends Command
         TemplateFinder $templateFinder,
         TemplateVariablesFactory $templateVariablesFactory,
         RectorRecipeProvider $rectorRecipeProvider,
-        PackageNamesProvider $packageNamesProvider,
-        NodeTypesProvider $nodeTypesProvider,
-        SetsListProvider $setsListProvider
+        RectorRecipeInteractiveProvider $rectorRecipeInteractiveProvider
     ) {
         parent::__construct();
 
@@ -109,9 +92,7 @@ final class GenerateCommand extends Command
         $this->overrideGuard = $overrideGuard;
         $this->fileGenerator = $fileGenerator;
         $this->rectorRecipeProvider = $rectorRecipeProvider;
-        $this->packageNamesProvider = $packageNamesProvider;
-        $this->nodeTypesProvider = $nodeTypesProvider;
-        $this->setsListProvider = $setsListProvider;
+        $this->rectorRecipeInteractiveProvider = $rectorRecipeInteractiveProvider;
     }
 
     protected function configure(): void
@@ -222,149 +203,6 @@ final class GenerateCommand extends Command
             return $this->rectorRecipeProvider->provide();
         }
 
-        return $this->prepareRectorRecipe();
-    }
-
-    private function prepareRectorRecipe(): RectorRecipe
-    {
-        $rectorRecipe = new RectorRecipe(
-            $this->askForPackageName(),
-            $this->askForRectorName(),
-            $this->askForNodeTypes(),
-            $this->askForRectorDescription(),
-            $this->getExampleCodeBefore(),
-            $this->getExampleCodeAfter(),
-        );
-        $rectorRecipe->setResources($this->askForResources());
-
-        $set = $this->askForSet();
-        if ($set !== null) {
-            $rectorRecipe->setSet($set);
-        }
-
-        return $rectorRecipe;
-    }
-
-    private function askForPackageName(): string
-    {
-        $question = new Question(sprintf(
-            'Package name for which Rector should be created (e.g. <fg=yellow>%s</>)',
-            'Naming'
-        ));
-        $question->setAutocompleterValues($this->packageNamesProvider->provide());
-
-        $packageName = $this->symfonyStyle->askQuestion($question);
-
-        return $packageName ?? $this->askForPackageName();
-    }
-
-    private function askForRectorName(): string
-    {
-        $question = sprintf(
-            'Class name of the Rector to create (e.g. <fg=yellow>%s</>)',
-            'RenameMethodCallRector',
-        );
-        $rectorName = $this->symfonyStyle->ask($question);
-
-        return $rectorName ?? $this->askForRectorName();
-    }
-
-    /**
-     * @return array<int, class-string>
-     */
-    private function askForNodeTypes(): array
-    {
-        $question = new ChoiceQuestion(sprintf(
-            'For what Nodes should the Rector be run (e.g. <fg=yellow>%s</>)',
-            'Expr/MethodCall',
-        ), $this->nodeTypesProvider->provide());
-        $question->setMultiselect(true);
-
-        $nodeTypes = $this->symfonyStyle->askQuestion($question);
-
-        $classes = [];
-        foreach ($nodeTypes as $nodeType) {
-            /** @var class-string $class */
-            $class = 'PhpParser\Node\\' . $nodeType;;
-            $classes[] = $class;
-        }
-
-        return $classes;
-    }
-
-    private function askForRectorDescription(): string
-    {
-        $description = $this->symfonyStyle->ask('Short description of new Rector');
-
-        return $description ?? $this->askForRectorDescription();
-    }
-
-    private function getExampleCodeBefore(): string
-    {
-        return <<<'CODE_SAMPLE'
-<?php
-
-class SomeClass
-{
-    public function run()
-    {
-        $this->something();
-    }
-}
-
-CODE_SAMPLE;
-    }
-
-    private function getExampleCodeAfter(): string
-    {
-        return <<<'CODE_SAMPLE'
-<?php
-
-class SomeClass
-{
-    public function run()
-    {
-        $this->somethingElse();
-    }
-}
-
-CODE_SAMPLE;
-    }
-
-    /**
-     * @return array<string>
-     */
-    private function askForResources(): array
-    {
-        $resources = [];
-
-        while (true) {
-            $question = sprintf(
-                'Link to resource that explains why the change is needed (e.g. <fg=yellow>%s</>)',
-                'https://github.com/symfony/symfony/blob/704c648ba53be38ef2b0105c97c6497744fef8d8/UPGRADE-6.0.md',
-            );
-            $resource = $this->symfonyStyle->ask($question);
-
-            if ($resource === null) {
-                break;
-            }
-
-            $resources[] = $resource;
-        }
-
-        return $resources;
-    }
-
-    private function askForSet(): ?string
-    {
-        $question = new Question(sprintf('Set to which Rector should be added (e.g. <fg=yellow>%s</>)', 'SYMFONY_52'));
-        $question->setAutocompleterValues($this->setsListProvider->provide());
-
-        $setName = $this->symfonyStyle->askQuestion($question);
-        if ($setName === null) {
-            return null;
-        }
-
-        return constant(SetList::class . $setName);
+        return $this->rectorRecipeInteractiveProvider->provide();
     }
 }
